Smart Pillbox/requirements.txt
awsiot
awscrt
RPi.GPIO
hx711
AWSIoTPythonSDK
